backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
reportlab==4.0.9
qrcode==7.4.2
pillow==10.2.0
pypdf==3.17.4
psutil==5.9.8
pydub==0.25.1
ffmpeg-python==0.2.0
openai==1.58.1
google-cloud-storage==2.14.0
# Force compatible versions for Python 3.11
typing-extensions>=4.5.0
pydantic>=2.0.0
httpx>=0.23.0
